spring-transaction-poc/failure-simulator/requirements.txt
requests
kafka-python